ubi_bao: divide config code
