+++ src/libvgmstream.vcxproj
@@ -187,6 +187,7 @@
     <ClInclude Include="meta\str_wav_streamfile.h" />
     <ClInclude Include="meta\txth_streamfile.h" />
     <ClInclude Include="meta\txtp.h" />
+    <ClInclude Include="meta\ubi_bao_config.h" />
     <ClInclude Include="meta\ubi_bao_streamfile.h" />
     <ClInclude Include="meta\ubi_ckd_cwav_streamfile.h" />
     <ClInclude Include="meta\ubi_lyn_streamfile.h" />
@@ -775,6 +776,7 @@
     <ClCompile Include="meta\txtp_process.c" />
     <ClCompile Include="meta\ubi_apm.c" />
     <ClCompile Include="meta\ubi_bao.c" />
+    <ClCompile Include="meta\ubi_bao_config.c" />
     <ClCompile Include="meta\ubi_ckd.c" />
     <ClCompile Include="meta\ubi_ckd_cwav.c" />
     <ClCompile Include="meta\ubi_hx.c" />
