+++ src/libvgmstream.vcxproj.filters
@@ -395,6 +395,9 @@
     <ClInclude Include="meta\txtp.h">
       <Filter>meta\Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="meta\ubi_bao_config.h">
+      <Filter>meta\Header Files</Filter>
+    </ClInclude>
     <ClInclude Include="meta\ubi_bao_streamfile.h">
       <Filter>meta\Header Files</Filter>
     </ClInclude>
@@ -2153,6 +2156,9 @@
       <Filter>meta\Source Files</Filter>
     </ClCompile>
     <ClCompile Include="meta\ubi_bao.c">
+      <Filter>meta\Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="meta\ubi_bao_config.c">
       <Filter>meta\Source Files</Filter>
     </ClCompile>
     <ClCompile Include="meta\ubi_ckd.c">
