+++ src/meta/ubi_bao.c
@@ -4,78 +4,15 @@
 #include "../util/endianness.h"
 #include "ubi_bao_streamfile.h"
 
+// BAO's massive confg and variations are handled here
+#include "ubi_bao_config.h"
+
 #define BAO_MIN_VERSION 0x1B
 #define BAO_MAX_VERSION 0x2A
 
 #define BAO_MAX_LAYER_COUNT 16      // arbitrary max
 #define BAO_MAX_CHAIN_COUNT 128     // POP:TFS goes up to ~100
 
-typedef enum { FILE_NONE = 0, FILE_FORGE, FILE_FORGE_b, FILE_FAT, FILE_DUNIA_CRC64 } ubi_bao_file_t;
-typedef enum { ARCHIVE_NONE = 0, ARCHIVE_ATOMIC, ARCHIVE_PK, ARCHIVE_SPK } ubi_bao_archive_t;
-typedef enum { TYPE_NONE = 0, TYPE_AUDIO, TYPE_LAYER, TYPE_SEQUENCE, TYPE_SILENCE } ubi_bao_type_t;
-typedef enum { CODEC_NONE = 0, UBI_IMA, RAW_PCM, RAW_PSX, RAW_XMA1, RAW_XMA2_OLD, RAW_XMA2_NEW, RAW_AT3, RAW_AT3_105, FMT_AT3, RAW_DSP, FMT_OGG } ubi_bao_codec_t;
-
-// config and offset for each field (since they move around depending on version)
-typedef struct {
-    bool big_endian;
-    bool allowed_types[16];
-    uint32_t version;
-
-    off_t bao_class;
-    size_t header_base_size;
-    size_t header_skip;
-
-    bool header_less_le_flag; // horrid but not sure what to do
-
-    off_t header_id;
-    off_t header_type;
-
-    off_t audio_stream_size;
-    off_t audio_stream_id;
-    off_t audio_external_flag;
-    off_t audio_loop_flag;
-    off_t audio_channels;
-    off_t audio_sample_rate;
-    off_t audio_num_samples;
-    off_t audio_num_samples2;
-    off_t audio_stream_type;
-    off_t audio_prefetch_size;
-    size_t audio_interleave;
-    off_t audio_cue_count;
-    off_t audio_cue_size;
-    bool audio_ignore_resource_size;
-    bool audio_fix_psx_samples;
-    int audio_external_and;
-    int audio_loop_and;
-
-    off_t sequence_sequence_loop;
-    off_t sequence_sequence_single;
-    off_t sequence_sequence_count;
-    off_t sequence_entry_number;
-    size_t sequence_entry_size;
-
-    off_t layer_layer_count;
-    off_t layer_external_flag;
-    off_t layer_stream_id;
-    off_t layer_stream_size;
-    off_t layer_prefetch_size;
-    off_t layer_extra_size;
-    off_t layer_cue_count;
-    off_t layer_cue_labels;
-    off_t layer_sample_rate;
-    off_t layer_channels;
-    off_t layer_stream_type;
-    off_t layer_num_samples;
-    size_t layer_entry_size;
-    bool layer_external_and;
-    bool layer_ignore_error;
-
-    off_t silence_duration_float;
-
-    ubi_bao_codec_t codec_map[16];
-    ubi_bao_file_t file;
-
-} ubi_bao_config_t;
 
 typedef struct {
     ubi_bao_archive_t archive;  // source format, which affects how related files are located
@@ -139,8 +76,6 @@ static STREAMFILE* setup_bao_streamfile(ubi_bao_header_t* bao, STREAMFILE* sf);
 static STREAMFILE* open_atomic_bao(ubi_bao_file_t file, uint32_t file_id, int is_stream, STREAMFILE* sf);
 static bool find_package_bao(uint32_t target_id, STREAMFILE* sf, off_t* p_offset, size_t* p_size);
 
-static bool ubi_bao_config_version(ubi_bao_config_t* cfg, STREAMFILE* sf, uint32_t version);
-static void ubi_bao_config_endian(ubi_bao_config_t* cfg, STREAMFILE* sf, off_t offset);
 static void build_readable_name(char* buf, size_t buf_size, ubi_bao_header_t* bao);
 
 
@@ -1553,447 +1488,4 @@ fail:
 
     VGM_LOG("UBI BAO: failed streamfile setup\n");
     return NULL;
-}
-
-/* ************************************************************************* */
-
-static void ubi_bao_config_endian(ubi_bao_config_t* cfg, STREAMFILE* sf, off_t offset) {
-
-    /* Detect endianness using the 'class' field (the 'header skip' field is LE in early
-     * versions, and was removed in later versions).
-     * This could be done once as all BAOs share endianness */
-
-    /* negate as fields looks like LE (0xN0000000) */
-    cfg->big_endian = !guess_endian32(offset + cfg->bao_class, sf);
-}
-
-
-static void config_bao_entry(ubi_bao_config_t* cfg, size_t header_base_size, size_t header_skip) {
-    cfg->header_base_size       = header_base_size;
-    cfg->header_skip            = header_skip;
-}
-
-/* audio header base */
-static void config_bao_audio_b(ubi_bao_config_t* cfg, off_t stream_size, off_t stream_id, off_t external_flag, off_t loop_flag, int external_and, int loop_and) {
-    cfg->audio_stream_size          = stream_size;
-    cfg->audio_stream_id            = stream_id;
-    cfg->audio_external_flag        = external_flag;
-    cfg->audio_loop_flag            = loop_flag;
-    cfg->audio_external_and         = external_and;
-    cfg->audio_loop_and             = loop_and;
-}
-/* audio header main */
-static void config_bao_audio_m(ubi_bao_config_t* cfg, off_t channels, off_t sample_rate, off_t num_samples, off_t num_samples2, off_t stream_type, off_t prefetch_size) {
-    cfg->audio_channels             = channels;
-    cfg->audio_sample_rate          = sample_rate;
-    cfg->audio_num_samples          = num_samples;
-    cfg->audio_num_samples2         = num_samples2;
-    cfg->audio_stream_type          = stream_type;
-  //cfg->audio_cue_count            = cue_count;
-  //cfg->audio_cue_labels           = cue_labels;
-    cfg->audio_prefetch_size        = prefetch_size;
-}
-
-static void config_bao_audio_c(ubi_bao_config_t* cfg, off_t cue_count, off_t cue_size) {
-    /* audio header extra */
-    cfg->audio_cue_count            = cue_count;
-    cfg->audio_cue_size             = cue_size;
-}
-
-static void config_bao_sequence(ubi_bao_config_t* cfg, off_t sequence_count, off_t sequence_single, off_t sequence_loop, off_t entry_size) {
-    /* sequence header and chain table */
-    cfg->sequence_sequence_count    = sequence_count;
-    cfg->sequence_sequence_single   = sequence_single;
-    cfg->sequence_sequence_loop     = sequence_loop;
-    cfg->sequence_entry_size        = entry_size;
-    cfg->sequence_entry_number      = 0x00;
-}
-
-static void config_bao_layer_m(ubi_bao_config_t* cfg, off_t stream_id, off_t layer_count, off_t external_flag, off_t stream_size, off_t extra_size, off_t prefetch_size, off_t cue_count, off_t cue_labels, int external_and) {
-    /* layer header in the main part */
-    cfg->layer_stream_id            = stream_id;
-    cfg->layer_layer_count          = layer_count;
-    cfg->layer_external_flag        = external_flag;
-    cfg->layer_stream_size          = stream_size;
-    cfg->layer_extra_size           = extra_size;
-    cfg->layer_prefetch_size        = prefetch_size;
-    cfg->layer_cue_count            = cue_count;
-    cfg->layer_cue_labels           = cue_labels;
-    cfg->layer_external_and         = external_and;
-}
-static void config_bao_layer_e(ubi_bao_config_t* cfg, off_t entry_size, off_t sample_rate, off_t channels, off_t stream_type, off_t num_samples) {
-    /* layer sub-headers in extra table */
-    cfg->layer_entry_size           = entry_size;
-    cfg->layer_sample_rate          = sample_rate;
-    cfg->layer_channels             = channels;
-    cfg->layer_stream_type          = stream_type;
-    cfg->layer_num_samples          = num_samples;
-}
-
-static void config_bao_silence_f(ubi_bao_config_t* cfg, off_t duration) {
-    /* silence headers in float value */
-    cfg->silence_duration_float     = duration;
-}
-
-
-static bool ubi_bao_config_version(ubi_bao_config_t* cfg, STREAMFILE* sf, uint32_t version) {
-
-    /* Ubi BAO evolved from Ubi SB and are conceptually quite similar, see that first.
-     *
-     * BAOs (binary audio objects) always start with:
-     * - 0x00(1): format (meaning defined by mode)
-     * - 0x01(3): 8b*3 version, major/minor/release (numbering continues from .sb0/sm0)
-     * - 0x04+: mini header (varies with version, see parse_header)
-     *
-     * Then are divided into "classes":
-     * - 0x10000000: event (links by id to another event or header BAO, may set volume/reverb/filters/etc)
-     * - 0x20000000: header
-     * - 0x30000000: memory audio (in .pk/.bao)
-     * - 0x40000000: project info
-     * - 0x50000000: stream audio (in .spk/.sbao)
-     * - 0x60000000: unused?
-     * - 0x70000000: info? has a count+table of id-things
-     * - 0x80000000: unknown (some floats?)
-     * - 0x90000000: unknown (some kind of command config?), rare [Ghost Recon Future Soldier (PC), Drawsome (Wii)]
-     * Class 1/2/3 are roughly equivalent to Ubi SB's section1/2/3, and class 4 is
-     * basically .spN project files.
-     *
-     * The project BAO (usually with special id 0x7FFFFFFF or 0x40000000) has version,
-     * filenames (not complete) and current mode, "PACKAGE" (pk, index + BAOs with
-     * external BAOs) or "ATOMIC" (file, separate BAOs).
-     *
-     * We want header classes, also similar to SB types:
-     * - 01: single audio (samples, channels, bitrate, samples+size, etc)
-     * - 02: play chain with config? (ex. silence + audio, or rarely audio 2ch intro + layer 4ch body)
-     * - 03: unknown chain
-     * - 04: random (count, etc) + BAO IDs and float probability to play
-     * - 05: sequence (count, etc) + BAO IDs and unknown data
-     * - 06: layer (count, etc) + layer headers
-     * - 07: unknown chain
-     * - 08: silence (duration, etc)
-     * - 09: silence with config? (channels, sample rate, etc), extremely rare [Shaun White Skateboarding (Wii)]
-     *
-     * Right after base BAO size is the extra table for that BAO (what sectionX had, plus
-     * extra crap like cue-like labels, even for type 0x01).
-     *
-     * Just to throw us off, the base BAO size may add +0x04 (with a field value of 0/-1) on
-     * some game versions/platforms (PC/Wii only?). Doesn't look like there is a header field
-     * (comparing many BAOs from different platforms of the same games) so it's autodetected.
-     *
-     * Most types + tables are pretty much the same as SB (with config styles ported straight) but
-     * now can "prefetch" part of the data (signaled by a size in the header, or perhaps a flag but
-     * looks too erratic). The header points to a external/stream ID, and with prefetch enabled part
-     * of the audio is in an internal/memory ID, and must join both during reads to get the full
-     * stream. Prefetch may be used in some platforms of a game only (ex. AC1 PC does while PS3
-     * doesn't, while Scott Pilgrim always does)
-     */
-
-    cfg->version = version;
-
-    cfg->allowed_types[0x01] = true;
-    cfg->allowed_types[0x05] = true;
-    cfg->allowed_types[0x06] = true;
-
-    /* absolute */
-    cfg->bao_class      = 0x20;
-
-    /* relative to header_skip */
-    cfg->header_id      = 0x00;
-    cfg->header_type    = 0x04;
-
-
-    /* 2 configs with same ID, autodetect */
-    if (cfg->version == 0x00220015) {
-        off_t header_size = 0x40 + read_u32le(0x04, sf); /* first is always LE */
-
-        /* next BAO uses machine endianness, entry should always exist
-         * (maybe should use project BAO to detect?) */
-        if (guess_endian32(header_size + 0x04, sf)) {
-            cfg->version |= 0xFF00; /* signal Wii=BE, but don't modify original */
-        }
-    }
-
-#if 0
-    if (cfg->version & 0xFFFF0000 >= 0x002B0000) {
-        /* 0x04(10): GUID
-         * 0x14: class
-         * 0x18: config/version? (0x02)
-         * 0x1c: fixed value per type? (0x17CE46D9 = header)
-         * (header skip)
-         * 0x20: header_id
-         * 0x34: header_type?
-         */
-        cfg->bao_class      = 0x14;
-        cfg->header_type    = 0x14; //from header_skip
-    }
-#endif
-
-    /* config per version (all offset from header_skip defined below) */
-    switch(cfg->version) {
-
-        case 0x001B0100: // Assassin's Creed (PS3/X360/PC)-atomic-forge
-      //case 0x001B0100: // My Fitness Coach (Wii)-atomic-forge
-      //case 0x001B0100: // Your Shape featuring Jenny McCarthy (Wii)-atomic-forge
-            config_bao_entry(cfg, 0xA4, 0x28); // PC: 0xA8, PS3/X360: 0xA4
-
-            config_bao_audio_b(cfg, 0x08, 0x1c, 0x28, 0x34, 1, 1); // 0x2c: prefetch flag?
-            config_bao_audio_m(cfg, 0x44, 0x48, 0x50, 0x58, 0x64, 0x74);
-            cfg->audio_interleave = 0x10;
-            cfg->audio_fix_psx_samples = true;
-
-            config_bao_sequence(cfg, 0x2c, 0x20, 0x1c, 0x14);
-
-            config_bao_layer_m(cfg, 0x4c, 0x20, 0x2c, 0x44, 0x00, 0x50, 0x00, 0x00, 1); // stream size: 0x48?
-            config_bao_layer_e(cfg, 0x30, 0x00, 0x04, 0x08, 0x10);
-
-            config_bao_silence_f(cfg, 0x1c);
-
-            cfg->codec_map[0x00] = RAW_XMA1;
-            cfg->codec_map[0x02] = RAW_PSX;
-            cfg->codec_map[0x03] = UBI_IMA;
-            cfg->codec_map[0x04] = FMT_OGG;
-            cfg->codec_map[0x05] = RAW_XMA1; // same but streamed?
-            cfg->codec_map[0x07] = RAW_AT3_105;
-
-            cfg->file = FILE_FORGE;
-            return true;
-
-        case 0x001B0200: // Beowulf (PS3/X360)-atomic-bin+fat
-            config_bao_entry(cfg, 0xA0, 0x24);
-
-            config_bao_audio_b(cfg, 0x08, 0x1c, 0x28, 0x34, 1, 1); // 0x2c: prefetch flag?
-            config_bao_audio_m(cfg, 0x44, 0x48, 0x50, 0x58, 0x64, 0x74);
-            cfg->audio_interleave = 0x10;
-            cfg->audio_fix_psx_samples = true;
-
-            config_bao_sequence(cfg, 0x2c, 0x20, 0x1c, 0x14);
-
-            config_bao_layer_m(cfg, 0x4c, 0x20, 0x2c, 0x44, 0x00, 0x50, 0x00, 0x00, 1); // stream size: 0x48?
-            config_bao_layer_e(cfg, 0x30, 0x00, 0x04, 0x08, 0x10);
-
-            config_bao_silence_f(cfg, 0x1c);
-
-            cfg->codec_map[0x00] = RAW_XMA1;
-            cfg->codec_map[0x02] = RAW_PSX;
-            cfg->codec_map[0x03] = UBI_IMA;
-            cfg->codec_map[0x04] = FMT_OGG;
-            cfg->codec_map[0x07] = RAW_AT3_105;
-
-            cfg->file = FILE_FAT;
-
-            return true;
-
-        case 0x001F0008: // Rayman Raving Rabbids: TV Party (Wii)-package
-        case 0x001F0010: // Prince of Persia 2008 (PC/PS3/X360)-atomic-forge, Far Cry 2 (PS3)-atomic-dunia?
-        case 0x001F0011: // Naruto: The Broken Bond (X360)-package
-        case 0x0021000C: // Splinter Cell: Conviction (E3 2009 Demo)(X360)-package
-        case 0x0022000D: // Just Dance (Wii)-package, We Dare (PS3/Wii)-package
-        case 0x0022FF15: // James Cameron's Avatar: The Game (Wii)-package
-        case 0x0022001B: // Prince of Persia: The Forgotten Sands (Wii)-package
-            config_bao_entry(cfg, 0xA4, 0x28); // PC/Wii: 0xA8
-
-            config_bao_audio_b(cfg, 0x08, 0x1c, 0x28, 0x34, 1, 1);
-            config_bao_audio_m(cfg, 0x44, 0x4c, 0x54, 0x5c, 0x64, 0x74); // cues: 0x68, 0x6c
-
-            config_bao_sequence(cfg, 0x2c, 0x20, 0x1c, 0x14);
-
-            config_bao_layer_m(cfg, 0x00, 0x20, 0x2c, 0x44, 0x4c, 0x50, 0x54, 0x58, 1); // 0x1c: id-like, 0x3c: prefetch flag?
-            config_bao_layer_e(cfg, 0x28, 0x00, 0x04, 0x08, 0x10);
-
-            config_bao_silence_f(cfg, 0x1c);
-
-            cfg->codec_map[0x01] = RAW_PCM;
-            cfg->codec_map[0x03] = UBI_IMA;
-            cfg->codec_map[0x04] = FMT_OGG;
-            cfg->codec_map[0x05] = RAW_XMA1;
-            cfg->codec_map[0x07] = RAW_AT3_105;
-            cfg->codec_map[0x09] = RAW_DSP;
-
-            cfg->file = FILE_FORGE_b;
-
-            if (cfg->version == 0x0022000D) // Just Dance (Wii) oddity
-                cfg->audio_ignore_resource_size = true;
-            if (cfg->version == 0x0022000D) // We Dare (Wii)
-                config_bao_audio_c(cfg, 0x68, 0x78);
-
-            return true;
-
-        case 0x00220015: // James Cameron's Avatar: The Game (PSP)-package
-        case 0x0022001E: // Prince of Persia: The Forgotten Sands (PSP)-package
-            config_bao_entry(cfg, 0x84, 0x28); // PSP: 0x84
-
-            config_bao_audio_b(cfg, 0x08, 0x1c, 0x20, 0x20, (1 << 2), (1 << 5)); // (1 << 4): prefetch flag?
-            config_bao_audio_m(cfg, 0x28, 0x30, 0x38, 0x40, 0x48, 0x58);
-
-            config_bao_layer_m(cfg, 0x00, 0x20, 0x24, 0x34, 0x3c, 0x40, 0x00, 0x00, (1 << 2)); // 0x1c: id-like
-            config_bao_layer_e(cfg, 0x28, 0x00, 0x04, 0x08, 0x10);
-
-            cfg->codec_map[0x06] = RAW_PSX;
-            cfg->codec_map[0x07] = FMT_AT3;
-
-            return true;
-
-        case 0x00230008: // Splinter Cell: Conviction (X360/PC)-package
-            config_bao_entry(cfg, 0xB4, 0x28); // PC: 0xB8, X360: 0xB4
-
-            config_bao_audio_b(cfg, 0x08, 0x24, 0x38, 0x44, 1, 1);
-            config_bao_audio_m(cfg, 0x54, 0x5c, 0x64, 0x6c, 0x74, 0x84);
-
-            config_bao_sequence(cfg, 0x34, 0x28, 0x24, 0x14);
-
-            config_bao_layer_m(cfg, 0x00, 0x28, 0x3c, 0x54, 0x5c, 0x00 /*0x60?*/, 0x00, 0x00, 1); // 0x24: id-like
-            config_bao_layer_e(cfg, 0x30, 0x00, 0x04, 0x08, 0x18);
-            cfg->layer_ignore_error = true; //TODO: last sfx layer (bass) may have smaller sample rate
-
-            cfg->codec_map[0x01] = RAW_PCM;
-            cfg->codec_map[0x02] = UBI_IMA;
-            cfg->codec_map[0x03] = FMT_OGG;
-            cfg->codec_map[0x04] = RAW_XMA2_OLD;
-
-            return true;
-
-        case 0x00250108: // Scott Pilgrim vs the World (PS3/X360)-package
-        case 0x0025010A: // Prince of Persia: The Forgotten Sands (PS3/X360)-atomic-forge
-        case 0x00250119: // Shaun White Skateboarding (Wii)-package
-        case 0x0025011D: // Shaun White Skateboarding (PC/PS3)-atomic-forge
-            config_bao_entry(cfg, 0xB4, 0x28); // PC: 0xB0, PS3/X360: 0xB4, Wii: 0xB8
-
-            if (cfg->version == 0x0025011D)
-                cfg->header_less_le_flag = true;
-
-            config_bao_audio_b(cfg, 0x08, 0x24, 0x2c, 0x38, 1, 1);
-            config_bao_audio_m(cfg, 0x48, 0x50, 0x58, 0x60, 0x68, 0x78);
-            cfg->audio_interleave = 0x10;
-
-            config_bao_sequence(cfg, 0x34, 0x28, 0x24, 0x14);
-
-            config_bao_layer_m(cfg, 0x00, 0x28, 0x30, 0x48, 0x50, 0x54, 0x58, 0x5c, 1); // 0x24: id-like
-            config_bao_layer_e(cfg, 0x30, 0x00, 0x04, 0x08, 0x18);
-            //TODO: some SPvsTW layers look like should loop (0x30 flag?)
-            //TODO: some POP layers have different sample rates (ambience)
-
-            config_bao_silence_f(cfg, 0x24);
-
-            cfg->codec_map[0x01] = RAW_PCM;
-            cfg->codec_map[0x02] = UBI_IMA;
-            cfg->codec_map[0x03] = FMT_OGG;
-            cfg->codec_map[0x04] = RAW_XMA2_NEW;
-            cfg->codec_map[0x05] = RAW_PSX;
-            cfg->codec_map[0x06] = RAW_AT3;
-            if (cfg->version == 0x0025010A) // no apparent flag
-                cfg->codec_map[0x06] = RAW_AT3_105;
-
-            cfg->file = FILE_FORGE_b;
-
-            return true;
-
-        case 0x00260000: // Michael Jackson: The Experience (X360)-package
-            config_bao_entry(cfg, 0xB8, 0x28);
-
-            config_bao_audio_b(cfg, 0x08, 0x28, 0x30, 0x3c, 1, 1); //loop?
-            config_bao_audio_m(cfg, 0x4c, 0x54, 0x5c, 0x64, 0x6c, 0x7c);
-
-            config_bao_layer_m(cfg, 0x00, 0x2c, 0x34,  0x4c, 0x54, 0x58,  0x00, 0x00, 1);
-            config_bao_layer_e(cfg, 0x34, 0x00, 0x04, 0x08, 0x1c);
-
-            cfg->codec_map[0x03] = FMT_OGG;
-            cfg->codec_map[0x04] = RAW_XMA2_NEW;
-
-            cfg->audio_ignore_resource_size = true; // leave_me_alone.pk
-
-            return true;
-
-        case 0x00270102: /* Drawsome (Wii)-package */
-            config_bao_entry(cfg, 0xAC, 0x28);
-
-            config_bao_audio_b(cfg, 0x08, 0x28, 0x2c, 0x38, 1, 1);
-            config_bao_audio_m(cfg, 0x44, 0x4c, 0x54, 0x5c, 0x64, 0x70);
-
-            config_bao_sequence(cfg, 0x38, 0x2c, 0x28, 0x14);
-
-            cfg->codec_map[0x02] = UBI_IMA;
-
-            return true;
-
-        case 0x00280303: // Tom Clancy's Ghost Recon Future Soldier (PC/PS3)-package
-            config_bao_entry(cfg, 0xBC, 0x28); // PC/PS3: 0xBC
-
-            config_bao_audio_b(cfg, 0x08, 0x38, 0x3c, 0x48, 1, 1);
-            config_bao_audio_m(cfg, 0x54, 0x5c, 0x64, 0x6c, 0x74, 0x80);
-
-            config_bao_sequence(cfg, 0x48, 0x3c, 0x38, 0x14);
-
-            config_bao_layer_m(cfg, 0x00, 0x3c, 0x44, 0x58, 0x60, 0x64, 0x00, 0x00, 1);
-            config_bao_layer_e(cfg, 0x2c, 0x00, 0x04, 0x08, 0x1c);
-
-            cfg->layer_ignore_error = true; //TODO: some layer sample rates don't match
-            //TODO: some files have strange prefetch+stream of same size (2 segments?), ex. CEND_30_VOX.lpk
-
-            config_bao_silence_f(cfg, 0x38);
-
-            cfg->codec_map[0x01] = RAW_PCM;
-            cfg->codec_map[0x02] = UBI_IMA; // v6
-            cfg->codec_map[0x04] = FMT_OGG;
-            cfg->codec_map[0x07] = RAW_AT3; //TODO: some layers use AT3_105
-
-            cfg->file = FILE_FORGE_b;
-
-            return true;
-
-        case 0x001C0000: // Lost: Via Domus (PS3)-atomic-gear
-            /* same as 0x001B0100 except:
-             * - base 0xA0, skip 0x24, name style %08x.bao (not .sbao?) */
-        case 0x001D0A00: // Shaun White Snowboarding (PSP)-atomic-opal
-        case 0x00220017: // Avatar (PS3)-atomic/spk
-        case 0x00220018: // Avatar (PS3)-atomic/spk
-        case 0x00260102: // Prince of Persia Trilogy HD (PS3)-package-gear
-            /* similar to 0x00250108 but most values are moved +4
-             * - base 0xB8, skip 0x28 */
-        case 0x00280306: // Far Cry 3: Blood Dragon (X360)-atomic-hashed
-        case 0x00290106: // Splinter Cell: Blacklist (PS3)-atomic-gear
-            /* quite different, lots of flags and random values
-             * - base varies per type (0xF0=audio), skip 0x20
-             * - 0x74: type, 0x78: channels, 0x7c: sample rate, 0x80: num_samples
-             * - 0x94: stream id? 0x9C: extra size */
-        case 0x002A0300: // Watch Dogs (Wii U)
-            /* similar to SC:B */
-      //case 0x002B0000: // Far Cry 4 (multi)-spk-dunia
-        case 0x002B0100: // Far Cry 4 (multi)-spk-dunia
-        #if 0
-            config_bao_entry(cfg, 0xD8, 0x20);
-
-            config_bao_audio_b(cfg, 0xA4?, 0x84, 0x40, 0x48?, 1, 1); //TODO: 54/74=alt header size? 94=alt stream size?
-            config_bao_audio_m(cfg, 0x68, 0x6c, 0x70/94, 0x70/94, 0x64, 0xA8);
-
-            /* debug info:
-                SAMPLE_INVALID = 0,
-                SAMPLE_PCM = 1,
-                SAMPLE_IMAADPCM = 2,
-                SAMPLE_IMADPCM_SEEKABLE_EVERYWHERE = 3,
-                SAMPLE_IMADPCM_SEEKABLE_ON_WAVE_MARKERS = 4,
-                SAMPLE_OGG = 5,
-                SAMPLE_XMA2 = 6,
-                SAMPLE_VAG = 7,
-                SAMPLE_MP3 = 8,
-                SAMPLE_DM = 9,
-                SAMPLE_ATRAC9 = 10,
-            */
-            cfg->codec_map[0x05] = FMT_OGG; // PC
-            cfg->codec_map[0x08] = FMT_MP3; // PS3
-            cfg->codec_map[0x0A] = RAW_AT9; // PS4
-
-            // stream/prefetch size go after extradata (must be detected + skipped)
-            // prefetch seems to be in .spk, streams in .sbao
-
-            // hashed with custom crc64 (see Gibbed.Dunia) from "soundbinary\%08x.%s" (w/ .spk/bao/sbao)
-            // ex. COMMON.FAT: 93D21F1037911836 > "soundbinary\2fffffff.spk"
-            cfg->file = FILE_DUNIA_CRC64;
-
-            break;
-        #endif
-
-        default: // others possibly using BAO: Watch_Dogs, Far Cry Primal
-            vgm_logi("UBI BAO: unknown BAO version %08x\n", cfg->version);
-            return false;
-    }
 }
